data-scripts/requirements.txt
requests==2.31.0
pandas==2.1.3
numpy==1.24.3
python-dotenv==1.0.0
pymongo==4.6.0
openpyxl==3.1.2
geopy==2.4.1
shapely==2.0.2


